Some general naming cleanup and file reorg.
The files that are no longer generated have had _gen removed
from their names. Some other minor changes to types and
function names and return types while still in alpha.

+++ clang/virtualfileoverlay.go
@@ -18,7 +18,7 @@ type VirtualFileOverlay struct {
 	Parameter options is reserved, always pass 0.
 */
 func NewVirtualFileOverlay() VirtualFileOverlay {
-	return VirtualFileOverlay{C.clang_VirtualFileOverlay_create(C.uint(0))}
+	return VirtualFileOverlay{C.clang_VirtualFileOverlay_create(0)}
 }
 
 // Map an absolute virtual file path to an absolute real one. The virtual path
